feat: move city bike info to own component
